Moved all files into `irc.ts` and `irc_test.ts` and fixed small issues
