--- irc/package.ts
@@ -1,2 +1,0 @@
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-export { assert, assertEqual, test } from "../testing/mod.ts";
