--- irc/test.ts
@@ -1,6 +1,0 @@
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-
-// runs all the tests in this module
-import "./mod_test.ts";
-import "./parser_test.ts";
-import "./user_test.ts";
